Struct gbmu::hardware::cpu::Cpu

pub struct Cpu {
    pub registers: Registers,
    state: State,
    memory: Memory,
}

Fields
registers: Registers

General-purpose registers, Program Counter and Stack Pointer

state: State

CPU state

memory: Memory

TODO

Implementations

impl Cpu

pub fn fetch(&self, address: u16) -> Result<(Opcode, u16), String>

Returns the Opcode enum matching the opcode read from memory If the opcode is 0xCB, the next byte is matched against CB-prefixed opcodes

fn fetch_cb(&self, address: u16) -> Opcode

impl Cpu

pub fn new(memory: Memory) -> Self

Initializes CPU with default values

fn read_imm8(&mut self) -> u8

Reads from the 8-bit immediate value from Program Counter.
Increments the Program Counter by 1.

fn read_imm16(&mut self) -> u16

Reads from the 16-bit immediate value from Program Counter.
Increments the Program Counter by 2.

fn decode(opcode: Opcode) -> Instruction

Returns the Instruction matching opcode.
If the opcode does not match any instructions, the function panics.

pub fn print_register(&self, register: Register8) -> String

pub fn step(&mut self)

TODO

fn execute(&mut self, instruction: Instruction)

fn load8(&mut self, destination: Operand8, source: Operand8)

Loads the value represented by source into destination.
This function calls get_operand8 to read the value from source and load_u8 to load it into destination.
Flag Register is not affected.

fn load_u8(&mut self, destination: Operand8, data: u8)

Loads the 8-bit data into destination.
If destination is either the 8-bit register A, B, C, D, E, H, L or an address (represented by Addr(at)), data is loaded. Otherwise, the function panics.

fn load8dec(&mut self, destination: Operand8, source: Operand8)

Loads the value of source into destination and decrements the value of the 16-bit register HL.

fn load8inc(&mut self, destination: Operand8, source: Operand8)

Loads the value of source into destination and increments the value of the 16-bit register HL.

fn load16(&mut self, destination: Operand16, source: Operand16)

Loads the 16-bit source into destination.
If destination is either the 16-bit register BC, DE, HL, SP or an address (represented by Addr(at)), data is loaded. Otherwise, the function panics.

fn push(&mut self, source: Operand16)

Pushes to the stack memory, the 16-bit source.
Before and after writing to memory, the stack pointer sp is decremented.
If source is not one of the 16-bit registers AF, BC, DE or HL, the function Registers::get_register16() panics.

fn pop(&mut self, target: Operand16)

Pops to the 16-bit register target, 16-bit of data pointed to by stack pointer sp.
After reading the stack memory, sp is incremented by 2.
If Operand16 is not a 16-bit register (AF, BC, DE, or HL), the function Registers::get_register16() panics.

fn cpl(&mut self)

Flips all the bits in the 8-bit register A.
Flag Register is updated as follows:
Z: Not affected
N: Set
H: Set
C: Not affected

fn daa(&mut self)

The DAA instruction adjusts the result of a binary addition or substraction to obtain the Binary Coded Decimal representation. Flag Register is updated as follows: Z: Set if the result is 0, otherwise reset N: Not affected H: Reset C: Set if there is carry from bit7, otherwise reset

fn dec8(&mut self, target: Operand8)

Decrements the content of target by 1.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Set
H: Set if there is a carry from bit3, otherwise reset
C: Not affected

fn inc8(&mut self, target: Operand8)

Increments data represented by target by 1.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Set if there is a carry from bit3, otherwise reset
C: Not affected

fn cp(&mut self, source: Operand8)

Subtracts source from the 8-bit register A without updating the content of A.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Set
H: Set if there is a carry from bit3, otherwise reset
C: Set if there is a carry from bit7, otherwise reset

fn or(&mut self, source: Operand8)

Performs a bitwise operation OR between 8-bit register A and source, and stores the result back into A.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Reset

fn xor(&mut self, source: Operand8)

Performs a bitwise operation XOR between 8-bit register A and source, and stores the result back into A. Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Reset

fn and(&mut self, source: Operand8)

Performs a bitwise AND operation between the 8-bit register A and source, and stores the result back into A.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Set
C: Reset

fn add_u8_to_a(&mut self, data: u8)

Adds data to the 8-bit register A and stores the result back into A.
FlagRegister is updated as follows:
Z: Set if the result is 0, otherwise reset
H: Set if there is a carry from bit3, otherwise reset
N: Reset
C: Set if there is a carry from bit7, otherwise reset

fn add8(&mut self, source: Operand8)

Adds source to the 8-bit register A, and stores the result back into A.
Calls add_u8_to_a with the value returned by get_operand8.

fn adc(&mut self, source: Operand8)

Adds source and the carry flag to the 8-bit register A, and stores the result back into A.
Calls add_u8_to_a with the value returned by get_operand8 and the carry flag.

fn sub_u8(&mut self, data: u8) -> u8

Substracts the 8-bit data from the 8-bit register A and returns the result.
FlagRegister is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Set
H: Set if there is a carry from bit3, otherwise reset
C: Set if there is a carry from bit7, otherwise reset

fn sub(&mut self, source: Operand8)

Substracts source from the 8-bit register A and stores the result back into A.
Calls sub_u8 with the value returned by get_operand8.

fn sbc(&mut self, source: Operand8)

Substracts source and the carry flag from the 8-bit register A and stores the result back into A.
Calls sub_u8 with the value returned by get_operand8 and the carry flag.

fn load_hl(&mut self)

Loads the sum of SP and the 8-bit immediate value to the 16-bit register HL.
Flag Register is updated as follows:
Z: Reset
N: Reset
H: Set if there is a carry from bit 11, otherwise reset
C: Set if there is a carry from bit 15, otherwise reset

fn add_sp_dd(&mut self)

Add to Stack Pointer the 8-bit immediate value.
Flag Register is updated as follows:
Z: Reset
N: Reset
H: Set if there is a carry on bit7, otherwise reset
C: Set if there is a carry on bit15, otherwise reset

fn inc16(&mut self, target: Operand16)

Increments data represented by target by 1.
Flag Register is not updated.

fn dec16(&mut self, target: Operand16)

Decrements data represented by target by 1.
Flag Register is not updated.

fn add_hl_r16(&mut self, source: Operand16)

Adds source to 16-bit register HL.
Flag Register are updated as follows:
Z: Not affected
H: Set if there is a carry from bit11, otherwise reset
N: Reset
C: Set if there is a carry from bit5, otherwise reset

fn swap(&mut self, target: Operand8)

Swap the lower and higher nibbles of the value represented by target.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Reset

fn srl(&mut self, target: Operand8)

Rotates the content of target to the right.
Copies bit0 to the carry flag and resets bit7.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Set if bit0 is 1 before the rotation, otherwise reset

fn sra(&mut self, target: Operand8)

Rotates the content of target to the right.
Copies bit0 to the carry flag and preserves bit7.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Set if bit0 is 1 before the rotation, otherwise reset

fn sla(&mut self, target: Operand8)

Rotates the content of target to the left.
Copies bit7 into the carry flag and resets bit0.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Set if bit7 is 1 before the rotation, otherwise reset

fn rr(&mut self, target: Operand8)

Rotates the content of target to the right.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Set if bit0 is 1 before the rotation, otherwise reset

fn rrc(&mut self, target: Operand8)

Rotates the content of target to the right.
Wraps the truncated bit to the end of the resulting integer.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Set if bit0 is 1 before the rotation, otherwise reset

fn rl(&mut self, target: Operand8)

Rotates the content of target to the left.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Set if bit7 is 1 before the rotation, otherwise reset

fn rlc(&mut self, target: Operand8)

Rotates the content of target to the left.
Wraps the truncated bit to the begining of the resulting integer.
Flag Register is updated as follows:
Z: Set if the result is 0, otherwise reset
N: Reset
H: Reset
C: Set if bit7 is 1 before the rotation, otherwise reset

fn rrca(&mut self)

Rotates the content of the 8-bit register A to the right.
Places the content of bit0 both in the carry flag and bit 7.
Flag Register is updated as follows:
Z: Reset
N: Reset
H: Reset
C: Set if bit0 is 1 before the rotation, otherwise reset

fn rla(&mut self)

Rotates the content of the 8-bit register A to the left.
Flag Register is updated as follows:
Z: Reset
N: Reset
H: Reset
C: Set if bit7 is 1 before the rotation, otherwise reset

fn rlca(&mut self)

Rotates the content of the 8-bit register A to the left.
Places the content of bit7 both in the carry flag and bit 0.
Flag Register is updated as follows:
Z: Reset
N: Reset
H: Reset
C: Set if bit7 is 1 before the rotation

fn bit(&mut self, bit: Bit, target: Operand8)

Copies the bit specified by bit in target to the zero flag.
Flag Register is updated as follows:
Z: Set if the specified bit is zero, otherwise reset
N: Reset
H: Set
C: Not affected

fn set(&mut self, bit: Bit, target: Operand8)

Sets the bit specified by bit in target to 1.
Flag Register is not updated.

fn res(&mut self, bit: Bit, target: Operand8)

Resets the bit specified by bit in target to 0.
Flag Register is not updated.

fn ccf(&mut self)

Toggles the carry flag.
If carry flag is set, then reset it.
If carry flag is reset, then set it.
Flag Register is updated as follows:
Z: Not affected
N: Reset
H: Reset
C: Toggled

fn scf(&mut self)

Sets the carry flag.
Flag Register is updated as follows:
Z: Not affected
N: Reset
H: Reset
C: Toggled

fn halt(&mut self)

Stops the system clock and enters HALT mode.
HALT mode is canceled by an interrupt or a reset signal.
Although the system clock is stopped in this state, the oscillator circuit and LCD controller continue to operate.

fn nop(&mut self)

Does nothing.

fn stop(&mut self)

Stops both the system clock and the oscillator circuit.
Stop mode stops the LCD controller.
Stop mode is canceled by a reset signal.

fn di(&mut self)

Disables interrupts

fn ei(&mut self)

Enables interrupts

fn jp(&mut self, condition: Condition, source: Operand16)

Loads the value of source to the Program Counter if condition is true.
Flag Register is not updated.

fn jr(&mut self, condition: Condition)

Adds the 8-bit immediate value to the Program Counter if condition is true.
Flag Register is not updated.

fn call(&mut self, condition: Condition, source: Operand16)

If condition is true, pushes Program Counter of the next instruction on the stack and loads source onto Program Counter.
Flag Register is not updated.

fn ret(&mut self, condition: Condition)

If condition is true, pops from the memory stack the 16-bit value and loads that value onto Program Counter.
Flag Register is not updated.

fn reti(&mut self)

Pops the 16-bit value on the top of memory stack and loads it onto Program Counter.
Enables the Master Interrupt flag.

fn rst(&mut self, address: Page0)

Loads the Program Counter into the memory stack and loads the page0 memory address onto Program Counter.

fn get_address(&mut self, addr: At) -> u16

Returns the 16-bit data stored at address represented by addr.
At can be the 16-bit register HL, BC, or DE,
a 16-bit immediate data (Imm16) or
an 8-bit data (Imm8 or 8-bit register C).
When addr represents an 8-bit data, the returned value is obtained by setting the most significant byte to 0xFF and the least significant to the 8-bit data to form an address in the range 0xFF00-0xFFFF.

fn get_operand8(&mut self, operand: Operand8) -> u8

Returns the 8-bit data represented by operand.
Operand8 is either a 8-bit register (A, B, C, D, E, H, L),
an 8-bit immediate data (Imm8) or
an 8-bit data stored at location (Addr(at) where at represents the location).

fn get_operand16(&mut self, operand: Operand16) -> u16

Returns the 16-bit data represented by operand.
Operand16 is either a 16-bit register (AF, BC, DE, HL, SP), a 16-bit or 8-bit immediate data (Imm16 or Imm8) or, the 16-bit data stored at location (Addr(at) where at represents the location). When operand represents an 8-bit data, the returned value is obtained by setting the most significant byte to 0x00 and the least significant byte to the 8-bit data to form a 16-bit value in the range 0x0000-0x00FF.

fn format_instruction( &self, imm: Imm, mnemonic: &str, address: u16 ) -> (String, u16)

TODO

pub fn disassemble(&self, lines: u16, address: u16) -> Vec<String>

TODO